Refactor out Jito_utils/js
