+++ package.json
@@ -1,6 +1,6 @@
 {
 	"name": "jup-grid",
-	"version": "0.5.1",
+	"version": "0.5.2",
 	"description": "JupGrid",
 	"type": "module",
 	"main": "src/jupgrid.js",
